Getting the eastern hemisphere times working; fixed an issue where the system bus stopped working, probably due to some C# changes from .NET 4.0 to .NET 4.7
