+++ src/DatabaseLib/StartStopTime.cs
@@ -7,6 +7,8 @@ namespace gov.sandia.sld.common.db
         private DateTimeOffset? _Start;
         private DateTimeOffset? _End;
 
+        private static readonly string _plusReplacement = "__";
+
         public DateTimeOffset? Start { get { return _Start; } }
         public DateTimeOffset? End { get { return _End; } }
 
@@ -17,6 +19,11 @@ namespace gov.sandia.sld.common.db
 
         public StartStopTime(string start, string end)
         {
+            if(string.IsNullOrEmpty(start) == false)
+                start = start.Replace(_plusReplacement, "+");
+            if (string.IsNullOrEmpty(end) == false)
+                end = end.Replace(_plusReplacement, "+");
+
             DateTimeOffset temp = DateTimeOffset.MinValue;
             if (DateTimeOffset.TryParse(start, out temp))
                 _Start = temp;
